requirements.txt
# Psycopg2
Flask
psycopg2-binary